Moved functionality to helper to reduce duplication
