+++ tailormap-components/projects/core/src/lib/feature-form/form-tree/form-tree-helpers.ts
@@ -6,7 +6,6 @@ import { AttributeListFeature } from '../../shared/attribute-service/attribute-m
 import { FormFieldHelpers } from '../form-field/form-field-helpers';
 import { TreeModel } from '../../shared/tree/models/tree.model';
 import { LayerUtils } from '../../shared/layer-utils/layer-utils.service';
-import { FormCreatorHelpers } from '../form-creator/form-creator-helpers';
 
 export class FormTreeHelpers {
 
@@ -62,13 +61,7 @@ export class FormTreeHelpers {
   public static getFeatureValueForField(feat: Feature | AttributeListFeature, config: FormConfiguration,
                                         key: string = config.treeNodeColumn): string  {
     const attr: Attribute = config.fields.find(field => field.key === key);
-    let value = null;
-    if (FormHelpers.isGeneratedFeature(feat)){
-      const field = FormCreatorHelpers.getAttribute(feat, key);
-      value = field?.value;
-    }else{
-      value = feat[key];
-    }
+    let value = FormHelpers.getValue(feat, attr);
 
     if (attr.type === FormFieldType.DOMAIN) {
       attr.options.forEach(option => {
