+++ tailormap-components/projects/core/src/lib/feature-form/form/form-helpers.ts
@@ -1,5 +1,7 @@
 import { Feature } from '../../shared/generated';
 import { AttributeListFeature } from '../../shared/attribute-service/attribute-models';
+import { Attribute } from './form-models';
+import { FormCreatorHelpers } from '../form-creator/form-creator-helpers';
 
 export class FormHelpers {
 
@@ -31,6 +33,17 @@ export class FormHelpers {
 
   public static isGeneratedFeature(feat: Feature | AttributeListFeature): feat is Feature{
     return feat.attributes;
+  }
+
+  public static getValue(feat : Feature | AttributeListFeature, field: Attribute) :string{
+    let value = null;
+    if(FormHelpers.isGeneratedFeature(feat)){
+      const attr = FormCreatorHelpers.getAttribute(feat, field.key);
+      value = attr?.value;
+    }else{
+      value = feat[field.key] ? feat[field.key] : '';
+    }
+    return value;
   }
 
 }
